see description:
Changed structure of view classes. For different realization of view
components now used class that implements
TicTacToeViewComponentsProvider interface (just added).
TicTacToeViewImpl gets instance of this interface in constructor and
uses it to initialize members such as gameBoardView etc.
TicTacToeViewImpl class is former TicTacToeViewAbstract class. The class
has been renamed because it ceased to be abstract class.